+++ src/app/admin/dashboard/page.js
@@ -9,9 +9,8 @@ import Navbar from "@/components/tutorme/home/nav/teacherNavbar";
 
 // import { Inter } from 'next/font/google'
 // import { useState } from "react";
-
+import SideBar from "@/components/tutorme/home/admin/sideBar";
 import ScrollBar from "./scrollBar";
-import SideBar from "./sideBar";
 
 const AcceptRequest = () => {
   // Your base page content goes here
@@ -28,7 +27,7 @@ const AcceptRequest = () => {
 
       <div className="flex w-100 space-x-4 p-4">
         <div className="w-[20%] shadow-md rounded-lg">
-          <SideBar />
+          <SideBar/>
         </div>
         <div className="w-[80%]">
           <ScrollBar />
